Attempt to use Listener and Client
https://docs.python.org/3.7/library/multiprocessing.html#module-multiprocessing.connection

This worked for a single client, however, the underlying code for Client uses blocking
connections.

I started to copy code to allow for non-blocking connections, however, having to do so
nullifies the point of using exiting infrastructure, if we only have to heavily alter/augment it.

This commit will be reverted and kept for reference.

Signed-off-by: James Noss <jnoss@stsci.edu>